sender/requirements.txt
msal==1.28.0
requests==2.31.0
PyJWT==2.8.0
